requirements.txt
pycairo>=1.23.0